+++ controller/hangman_acp.php
@@ -1,9 +1,9 @@
 <?php
 /*
 *
-* @package Hangman v0.4.0
+* @package Hangman v0.7.0
 * @author Mike-on-Tour
-* @copyright (c) 2021 - 2022 Mike-on-Tour
+* @copyright (c) 2021 - 2023 Mike-on-Tour
 * @former author dmzx (www.dmzx-web.net)
 * @copyright (c) 2015 by dmzx (www.dmzx-web.net)
 * @license http://opensource.org/licenses/gpl-license.php GNU Public License
@@ -98,6 +98,7 @@ class hangman_acp
 			$this->config->set('mot_hangman_show_term', $this->request->variable('mot_hangman_show_term', 0));
 			$this->config->set('mot_hangman_enforce_term', $this->request->variable('mot_hangman_enforce_term', 0));
 			$this->config->set('mot_hangman_enforce_term_ratio', $this->request->variable('mot_hangman_enforce_term_ratio', 0));
+			$this->config->set('mot_hangman_display_online', $this->request->variable('mot_hangman_display_online', 0));
 			$this->config->set('mot_hangman_rows_per_page', $this->request->variable('mot_hangman_rows_per_page', 0));
 			$this->config->set('mot_hangman_lives', $this->request->variable('acp_hangman_lives', 0));
 			$this->config->set('mot_hangman_points_letter', $this->request->variable('acp_hangman_points_letter', 0));
@@ -105,6 +106,8 @@ class hangman_acp
 			$this->config->set('mot_hangman_points_win', $this->request->variable('acp_hangman_points_win', 0));
 			$this->config->set('mot_hangman_points_word', $this->request->variable('acp_hangman_points_word', 0));
 			$this->config->set('mot_hangman_evade_enable', $this->request->variable('mot_hangman_evade_enable', 0));
+			$this->config->set('mot_hangman_extra_points_enable', $this->request->variable('mot_hangman_extra_points_enable', 0));
+			$this->config->set('mot_hangman_extra_points', $this->request->variable('mot_hangman_extra_points', 0));
 			$this->config->set('mot_hangman_term_length', $this->request->variable('acp_hangman_term_length', 0));
 			$this->config->set('mot_hangman_punctuation_marks', $this->request->variable('acp_hangman_punctuation_marks', ''));
 
@@ -327,12 +330,15 @@ class hangman_acp
 			'ACP_MOT_HANGMAN_ENFORCE_TERM'			=> $this->config['mot_hangman_enforce_term'],
 			'ACP_MOT_HANGMAN_ENFORCE_TERM_RATIO'	=> $this->config['mot_hangman_enforce_term_ratio'],
 			'ACP_MOT_HANGMAN_ROWS_PER_PAGE'			=> $this->config['mot_hangman_rows_per_page'],
+			'ACP_MOT_HANGMAN_DISPLAY_ONLINE'		=> $this->config['mot_hangman_display_online'],
 			'ACP_MOT_HANGMAN_LIVES'					=> $this->config['mot_hangman_lives'],
 			'ACP_MOT_HANGMAN_POINTS_LETTER'			=> $this->config['mot_hangman_points_letter'],
 			'ACP_MOT_HANGMAN_POINTS_LOOSE'			=> $this->config['mot_hangman_points_loose'],
 			'ACP_MOT_HANGMAN_POINTS_WIN'			=> $this->config['mot_hangman_points_win'],
 			'ACP_MOT_HANGMAN_POINTS_WORD'			=> $this->config['mot_hangman_points_word'],
 			'ACP_MOT_HANGMAN_EVADE_ENABLE'			=> $this->config['mot_hangman_evade_enable'],
+			'MOT_HANGMAN_EXTRA_POINTS_ENABLE'		=> $this->config['mot_hangman_extra_points_enable'],
+			'MOT_HANGMAN_EXTRA_POINTS'				=> $this->config['mot_hangman_extra_points'],
 			'ACP_MOT_HANGMAN_TERM_LENGTH'			=> $this->config['mot_hangman_term_length'],
 			'ACP_MOT_HANGMAN_PUNCTUATION_MARKS'		=> $this->config['mot_hangman_punctuation_marks'],
 			'ACP_MOT_HANGMAN_IMPORT_OLD_TABLE'		=> $this->db_tools->sql_table_exists($this->old_hangman_words_table),