+++ language/de/info_acp_hangman.php
@@ -1,8 +1,8 @@
 <?php
 /**
 *
-* @package Hangman v0.5.0
-* @copyright (c) 2021 - 2022 Mike-on-Tour
+* @package Hangman v0.7.0
+* @copyright (c) 2021 - 2023 Mike-on-Tour
 * @license http://opensource.org/licenses/gpl-2.0.php GNU General Public License v2
 *
 */
@@ -33,30 +33,32 @@ if (empty($lang) || !is_array($lang))
 // in a url you again do not need to specify an order e.g., 'Click %sHERE%s' is fine
 
 $lang = array_merge($lang, [
+	'ACP_HANGMAN'								=> 'Hangman',
 	'ACP_MOT_HANGMAN'							=> 'Hangman',
+	'ACP_HANGMAN_SETTINGS'						=> 'Einstellungen',
 	'ACP_MOT_HANGMAN_SETTINGS'					=> 'Einstellungen',
 	// General settings
 	'ACP_MOT_HANGMAN_EXT_SETTINGS'				=> 'Programmeinstellungen',
 	'ACP_MOT_HANGMAN_AUTODELETE'				=> 'Benutzte Begriffe löschen?',
-	'ACP_MOT_HANGMAN_AUTODELETE_EXP'			=> 'Wenn du hier ´Ja´ auswählst, werden beim Spiel verwendete Suchbegriffe bei Beendigung des Spiels aus der Datenbank gelöscht.',
+	'ACP_MOT_HANGMAN_AUTODELETE_EXP'			=> 'Wenn aktiviert, werden beim Spiel verwendete Suchbegriffe bei Beendigung des Spiels aus der Datenbank gelöscht.',
 	'ACP_MOT_HANGMAN_CATEGORY_ENABLE'			=> 'Kategorien einschalten?',
-	'ACP_MOT_HANGMAN_CATEGORY_ENABLE_EXP'		=> 'Wenn du hier ´Ja´ auswählst, wird bei der Eingabe eines Suchbegriffes ein weiteres Eingabefeld angezeigt, in dem für
+	'ACP_MOT_HANGMAN_CATEGORY_ENABLE_EXP'		=> 'Wenn aktiviert, wird bei der Eingabe eines Suchbegriffes ein weiteres Eingabefeld angezeigt, in dem für
 													den eingegebenen Suchbegriff eine Kategorie angegeben werden kann.<br>
 													Diese Kategorie wird im Spiel über dem Feld mit dem zu suchenden Begriff angezeigt und dient so als kleine Hilfestellung.',
 	'ACP_MOT_HANGMAN_CATEGORY_ENFORCE'			=> 'Ausfüllen der Kategorie erzwingen?',
-	'ACP_MOT_HANGMAN_CATEGORY_ENFORCE_EXP'		=> 'Wenn du hier ´Ja´ auswählst, wird bei eingeschalteter ´Kategorie´ das Eingeben einer Kategorie erzwungen, das
+	'ACP_MOT_HANGMAN_CATEGORY_ENFORCE_EXP'		=> 'Wenn aktiviert, wird bei eingeschalteter ´Kategorie´ das Eingeben einer Kategorie erzwungen, das
 													Eingabeformular kann dann ohne eine Eintragung im Feld ´Kategorie´ nicht verlassen werden.',
 	'ACP_MOT_HANGMAN_SHOW_TERM'					=> 'Anzeige des Suchbegriffes bei Verlieren',
 	'ACP_MOT_HANGMAN_SHOW_TERM_EXP'				=> 'Wenn aktiviert, wird in der Meldung über das verlorene Spiel auch der Suchbegriff im Klartext angezeigt.',
 	'ACP_MOT_HANGMAN_ENFORCE_TERM'				=> 'Spieler mit zu wenigen Eingabepunkten blockieren',
-	'ACP_MOT_HANGMAN_ENFORCE_TERM_EXP'			=> 'Wenn du hier ´Ja´ auswählst, werden Spieler, bei denen das Verhältnis der Spielepunkte zu den Begriffseingabepunkten
+	'ACP_MOT_HANGMAN_ENFORCE_TERM_EXP'			=> 'Wenn aktiviert, werden Spieler, bei denen das Verhältnis der Spielepunkte zu den Begriffseingabepunkten
 													eine bestimmte Grenze überschreitet, aufgefordert vor einem neuen Spiel zunächst neue Suchbegriffe einzugeben.
 													So kannst du Spieler erziehen, die das Eingaben von Suchbegriffen vernachlässigen.<br>
 													Das Verhältnis Spielepunkte zu Begriffseingabepunkten kannst du in der nächsten Einstellung auswählen, sie
 													wird angezeigt sobald du hier ´Ja´ auswählst.',
 	'ACP_MOT_HANGMAN_ENFORCE_TERM_RATIO'		=> 'Verhältnis Spielepunkte zu Begriffseingabepunkten',
 	'ACP_MOT_HANGMAN_ENFORCE_TERM_RATIO_EXP'	=> 'Es handelt sich bei dieser Maßnahme prinzipiell um den Ausschluss von Mitgliedern vom Spiel, wenn sie zu
-													zu wenige neue Suchbegriffe eingeben; es wird deshalb empfohlen, dieses Verhältnis anfangs nicht zu niedrig
+													wenige neue Suchbegriffe eingeben; es wird deshalb empfohlen, dieses Verhältnis anfangs nicht zu niedrig
 													einzustellen, beispielsweise bedeutet der voreingestellte Wert von 40, dass ein Spieler mit 40 Begriffseingabepunkten
 													und mehr als 1.600 Spielepunkten so lange vom Spiel ausgeschlossen wird (er bekommt statt des Spiels einen
 													entsprechenden Hinweistext angezeigt) bis er so viele neue Suchbegriffe eingegeben hat, dass das Verhältnis
@@ -69,6 +71,9 @@ $lang = array_merge($lang, [
 													von 3,125, aufgerundet 4. Damit muss jeder Spieler für (fast) jeden verwendeten Begriff einen
 													neuen eingeben; dies erscheint als gesundes Verhältnis, wenn verwendete Begriffe gelöscht
 													werden, ist dies nicht der Fall, kann das Verhältnis auch deutlich höher gewählt werden.',
+	'ACP_MOT_HANGMAN_DISPLAY_ONLINE'			=> 'Anzeige der aktuell aktiven Hangman-Spieler',
+	'ACP_MOT_HANGMAN_DISPLAY_ONLINE_EXP'		=> 'Bei Aktivierung dieser Einstellung werden auf der Foren-Übersicht im Bereich „Wer ist online“ die Gesamtanzahl und die Namen
+													der Mitglieder angezeigt, die gerade eine Funktion von Hangman nutzen.',
 	'ACP_MOT_HANGMAN_ROWS_PER_PAGE'				=> 'Zeilen pro Tabellenseite',
 	'ACP_MOT_HANGMAN_ROWS_PER_PAGE_EXP'			=> 'Wähle hier die Anzahl der Zeilen, die pro Tabellenseite angezeigt werden soll.',
 	// Game settings
@@ -91,12 +96,17 @@ $lang = array_merge($lang, [
 	'ACP_MOT_HANGMAN_POINTS_WORD_EXP'			=> 'Anzahl der Punkte, die einem Spieler für das Erstellen eines Suchbegriffes auf seinem Spielekonto gutgeschrieben werden<br>
 													(ganze Zahl größer oder gleich Null)',
 	'ACP_MOT_HANGMAN_EVADE_ENABLE'				=> 'Verlassen eines Spieles als Verloren werten?',
-	'ACP_MOT_HANGMAN_EVADE_ENABLE_EXP'			=> 'Wenn du hier ´Ja´ auswählst, erhalten Spieler, die ein laufendes Spiel verlassen (z.B. durch Anklicken eines
+	'ACP_MOT_HANGMAN_EVADE_ENABLE_EXP'			=> 'Wenn aktiviert, erhalten Spieler, die ein laufendes Spiel verlassen (z.B. durch Anklicken eines
 													Links oder Neuladen der Seite) die ´Punktezahl bei Verlieren´ angerechnet.<br>
 													Ist diese Option eingeschaltet, werden Spieler über ein Popup-Fenster gefragt, ob sie das Spiel wirklich
 													verlassen wollen und erst bei Bejahen dieser Frage mit der ´Punktezahl bei Verlieren´ bestraft. Abhängig von
 													der Einstellung ´Benutzte Begriffe löschen´ wird der in diesem Spiel verwendete Suchbegriff ggf. gelöscht.<br>
 													Ist diese Option ausgeschaltet, wird die verlangte Seite ohne weitere Maßnahmen geladen.',
+	'MOT_HANGMAN_EXTRA_POINTS_ENABLE'			=> 'Extrapunkte bei fehlerfreiem Lösen',
+	'MOT_HANGMAN_EXTRA_POINTS_ENABLE_EXP'		=> 'Wenn du die Hangman-Spieler mit Extrapunkten für das Lösen ohne falsche Versuche belohnen möchtest, kannst du das hier
+													aktivieren; nach Aktivierung kannst du in einer weiteren, dann sichtbaren Einstellung die Anzahl der Extrapunkte eingeben.',
+	'MOT_HANGMAN_EXTRA_POINTS'					=> 'Extrapunkte für fehlerfreies Lösen',
+	'MOT_HANGMAN_EXTRA_POINTS_EXP'				=> 'Hier kannst du die Anzahl der Extrapunkte eingeben, Minimum ist ein Extrapunkt, Maximum sind 50 Extrapunkte.',
 	//Term settings
 	'ACP_MOT_HANGMAN_INPUT_SETTINGS'			=> 'Einstellungen für Suchbegriff',
 	'ACP_MOT_HANGMAN_INPUT_SETTINGS_EXP'		=> 'Hier kannst du Einstellungen für die Eingabe der Suchbegriffe vornehmen.',