+++ language/en/mot_hangman_ext_enable_error.php
@@ -23,5 +23,5 @@ $lang = array_merge($lang, [
 	'MOT_HANGMAN_EXT_NAME'						=> 'Hangman',
 	'MOT_HANGMAN_ERROR_EXTENSION_NOT_ENABLE'	=> 'The extension „%1$s“ can not be enabled. Please check whether the necessary requirements for this extension are satisfied.',
 	'MOT_HANGMAN_ERROR_MESSAGE_PHPBB_VERSION'	=> 'Minimum version of phpBB required is „%1$s“ but less than „%2$s“',
-	'MOT_HANGMAN_PHP_VERSION_ERROR'				=> 'Minimum version of PHP is „%1$s“',
+	'MOT_HANGMAN_PHP_VERSION_ERROR'				=> 'Minimum version of PHP is „%1$s“ but less than „%2$s“',
 ]);